requirements.txt
requests
ipwhois
dnspython
colorama
